Usuarios ya puede guardar empresa Producto ya funciona para crear y editar Mejora de diversos modelos

+++ src/usuarios/model/usuario.model.ts
@@ -25,14 +25,10 @@ export class Usuario {
     @JoinColumn({ name: 'EmpresaId' })
     Empresa: Empresa;
 
-    @Column('numeric',{
-        default: 0
-    })
+    @Column('int', { default: 0 })
     intentosConexion: number;
 
-    @Column('numeric',{
-        default: 0
-    })
+    @Column('int', { default: 0 })
     expiracionTiempo: number;
 
     @Column('boolean',{
